+++ Editor/OptimizerPlugin.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using Anatawa12.AvatarOptimizer.ndmf;
 using nadena.dev.ndmf;
 using nadena.dev.ndmf.builtin;
@@ -77,6 +77,7 @@ namespace Anatawa12.AvatarOptimizer.ndmf
                         .Then.Run(Processors.TraceAndOptimizes.AutoMergeSkinnedMesh.Instance)
                         .Then.Run(Processors.TraceAndOptimizes.FindUnusedObjects.Instance)
                         .Then.Run(Processors.TraceAndOptimizes.ConfigureRemoveZeroSizedPolygon.Instance)
+                        .Then.Run(Processors.TraceAndOptimizes.RemoveUnusedMaterialProperties.Instance)
                         .Then.Run(Processors.MergeBoneProcessor.Instance)
                         .Then.Run(Processors.RemoveZeroSizedPolygonProcessor.Instance)
                         .Then.Run(Processors.TraceAndOptimizes.OptimizeTexture.Instance)