+++ Runtime/TraceAndOptimize.cs
@@ -138,6 +138,8 @@ namespace Anatawa12.AvatarOptimizer
             public bool skipRemoveEmptySubMesh;
             [ToggleLeft]
             public bool skipAnyStateToEntryExit;
+            [ToggleLeft]
+            public bool skipRemoveMaterialUnusedProperties;
         }
     }
 }